OP-378: Small changes to get drivers working again.  Fix I2C port for IMU3000 and change convention so that 0 is success, < 0 is error.

+++ flight/PiOS/inc/pios_imu3000.h
@@ -120,7 +120,7 @@ struct pios_imu3000_data {
 extern void PIOS_IMU3000_Init(void);
 extern bool PIOS_IMU3000_NewDataAvailable(void);
 extern uint8_t PIOS_IMU3000_ReadGyros(struct pios_imu3000_data * data);
-extern uint8_t PIOS_IMU3000_ReadID();
+extern int32_t PIOS_IMU3000_ReadID();
 extern uint8_t PIOS_IMU3000_Test();
 
 #endif /* PIOS_IMU3000_H */
